+++ src/parser/inline/comment.ts
@@ -1,24 +1,24 @@
 import { CommentParser } from '../inline';
-import { eval } from '../../combinator/data/parser';
-import { union, some, validate, creator, surround, match, lazy } from '../../combinator';
+import { union, some, validate, creator, surround, open, close, match, lazy } from '../../combinator';
 import { inline } from '../inline';
 import { text, str } from '../source';
 import { html, defrag } from 'typed-dom';
 import { memoize } from 'spica/memoize';
-import { unshift, push, pop } from 'spica/array';
+import { unshift, push } from 'spica/array';
 
 export const comment: CommentParser = lazy(() => creator(validate('[#', match(
   /^(?=\[(#+)\s)/,
   memoize(
-  ([, fence], closer = new RegExp(String.raw`^\s+${fence}\]`)) =>
+  ([, fence]) =>
     surround(
-      str(/^\[(\S+)\s+(?!\1\])/),
-      union([some(inline, closer)]),
-      str(closer), true,
-      ([, bs = []], rest) => [[
-        html('span',
-          { class: 'comment' },
-          defrag(push(unshift([`[${fence} `], bs), [` ${fence}]`]))),
+      open(str(`[${fence}`), some(text, new RegExp(String.raw`^\s+${fence}\]|^\S`)), true),
+      union([some(inline, new RegExp(String.raw`^\s+${fence}\]`))]),
+      close(some(text, /^\S/), str(`${fence}]`)), true,
+      ([as, bs = [], cs], rest) => [[
+        html('span', { class: 'comment' }, [
+          html('input', { type: 'checkbox' }),
+          html('span', defrag(push(unshift(as, bs), cs))),
+        ]),
       ], rest],
-      ([as, bs = []], rest, context) => [unshift(pop(eval(some(text)(`${as[0]}!`, context))!)[0], bs), rest]),
+      ([as, bs = []], rest) => [unshift(as, bs), rest]),
   ([, fence]) => fence)))));